moving middleware around
